requirements.txt
pbr!=2.1.0,>=2.0.0 # Apache-2.0
pyyaml
